Refactor MSDeployArtifactHandler, extract common codes (#895)
