+++ azure-functions-maven-plugin/src/main/java/com/microsoft/azure/maven/function/Constants.java
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for
+ * license information.
+ */
+
+package com.microsoft.azure.maven.function;
+
+public class Constants {
+    public static final String ZIP_EXT = ".zip";
+    public static final String LOCAL_SETTINGS_FILE = "local.settings.json";
+    public static final String INTERNAL_STORAGE_KEY = "AzureWebJobsStorage";
+}
